# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Frases Célebres API",
        "description": "API sobre frases célebres y motivadoras en castellano. Contiene categoria y autores.",
        "contact": {
            "email": "esmeldyfm@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Acceso"
                ],
                "summary": "Registrar nuevo usuario",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nombre de usuario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Correo electrónico",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Contraseña",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario registrado correctamente"
                    },
                    "401": {
                        "description": "Error de validación de datos"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Acceso"
                ],
                "summary": "Login de usuarios",
                "description": "Login con datos de usuarios registrados previamente.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Correo electrónico",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Contraseña",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario logueado correctamente"
                    },
                    "401": {
                        "description": "No autorizado."
                    }
                }
            }
        },
        "/api/logout": {
            "get": {
                "tags": [
                    "Acceso"
                ],
                "summary": "Cerrar sessión de usuario",
                "operationId": "3f19b742e4ead13f81ba0d6c0d46ac1b",
                "responses": {
                    "200": {
                        "description": "Sessión cerrada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Acceso",
            "description": "Endpoints de acceso"
        },
        {
            "name": "Autores",
            "description": "Endpoints de autores"
        }
    ]
}